requirements.txt
Flask==3.0.2
google-generativeai==0.4.0
pillow==10.2.0
